third_party/WebKit/LayoutTests/crypto/subtle/rsassa-pkcs1-v1_5/sign-verify-expected.txt

Test importing an RSA key for RSASSA-PKCS1-v1_5.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing a public key...

Importing a private key...

Signing some text...
PASS bytesToHexString(new Uint8Array(signature)) is '0fd9a8aef4cc1876c0b762545336c6d1fb315ae16ae4b5e4bf34d384d8585ea7a01e76ea09ee7f7ee8d1c122e7dd15b7c94a573b2aa07203e8d13bc6fd16156cd8e5f0c15a15dccb62d152127fca09882fb53bc3e60ab586d15b95cf411e3aab4a1c231a7e91aab09ee3d4b13d11e97505ddff77683470da510ee76e8bd530c56a85f901626a5a710f716f113dfe9cf6c473ee16fa248aea3480a1033abe30f4c1243289a661e64d7818b55698280688097135968c6d4b029496d85cab2a67e4696737781f70e4392c7df71bbd6c92465947f029a1de48160aced11b5721b1cd25039fe2c16c2b38de73df3b9a83e3ea755fd0cfe51ca06b61fadf6d84677f95'

Verifying the signature...
PASS verificationResult is true

Verifying a bad signature...
PASS verificationResult is false
PASS successfullyParsed is true

TEST COMPLETE
